Move 'addProject', 'removeProject', 'handleProject' to separate files.
