+++ src/components/form/components/projects/utils/handleProject.js
@@ -1,0 +1,5 @@
+export const handleProject = (resumeData, setResumeData, e, index) => {
+  const newProjects = [...resumeData.projects];
+  newProjects[index][e.target.name] = e.target.value;
+  setResumeData({...resumeData, projects: newProjects});
+};
